requirements.txt
aiogram>=3.6
python-dotenv>=1.0
certifi>=2024.0.0